Rename: SATSFont.FontVariant to SATSFont.Weight
the values we use for this are:

- default -> Regular
- medium -> Medium
- emphasis -> SemiBold
- sats feeling -> BoldItalic

So i was not particularly happy with the variation name, even though the
sats feeling case has a `bold` and italic to the font type we use.

I felt overall that `Weight` was a better name than `FontVariant`

What do you think? Do you have a better suggestion?