Fix layout issue in TopBar
autolayout was complaining due to a conflicting constraint error, which
was due mostly due constraints that UIStackView makes.

Given that UIStackView is not doing much in this particular view I opted
to removed it from the internal implementation.

This doesn't change anything from the API or behavior of the component.