Add initial implementation of `SwiftUI` SATSButton styles.
Button styles are applied with the method

```swift
Button("Sample", action: {})
    .satsButton(.primary, size: .large)
```

There are some refinements to do for these styles. One of the bugs is that
the button "jumps" when showing the spinner, as it has a different size than
the text. Also, the spinner must be something we use in the app + the color
tint of it has to be refined