Use `@ObservedObject` instead of `@StateObject`
for the internal `ImageViewModel` instance.

This is clear when in the demo change the value of the image state we
want to show, if we have an `StateObject` that view model will be just
created once, which will ignore the new state passed onto
`CustomAsyncImage`.

Then, in order to make `CustomAsyncImage` react to external state
changes, we create a new `ImageViewModel` when needed with
`ObservedObject`.

This change is slightly risky as we may open the door for a potential
memory leak (as explained when apple introduced `@StateObject`) or it
may be just the right behavior. This is where our inexperience with
SwiftUI comes to play, so we can test this and fix it if necessary.

Since we don't require `@StateObject` anyore, this component is now
available for iOS 13 as well.