Add `IdentifiableList` Data Structure.
When using SwiftUI and lists, with values that can update their state
like a `isLoading` property. We wanted to use a `Set` as a backing data
structure, as an Array would mean that every time we want to mutate a
given value, we need to find its index first.

`Set` uses `Hashable` and `Equatable` conformances to determine if a
given element is present on the list.

In `SwiftUI` it's common to use the `Identifiable` for elements in a
list, so SwiftUI will update them correctly.

In general swiftUI required 2 properties to see if the UI needs to be
refreshed or not:
1. Is the value the same (by ID)
2. Is the value the same (by values)

For this reason, we cannot use a custom `Equatable` conformance for
values that have an identity which value changes.

Then I decided to create a simple structure that as a collection
fulfills the 2 SwiftUI requirements, `IdentifiableList`.

There are 2 caveats with this solution:

1. As in right now, we don't even conform to `Collection` or subtypes
as I need to understand better those protocols to implement the
conformance, something I may do in the future.

2. There are libraries like [swift-identified-collections](https://github.com/pointfreeco/swift-identified-collections)
that implement a full featured collection, but I didn't wanted to add
more dependencies.

So in that regard we can start by rolling out with our simple
implementation and evaluate then if we need a more feature-rich one in
the future.