Replace strings for enum values for theme names
These enum values work as IDs for the different themes.

Having a `ColorTheme` struct localizes the different values that compose a theme,
in opposition of having themes as a `enum` and then we have a bunch of
`switch` statements for each property.

Having the `name` as a `enum` is a nice compromise where given a name
we can ensure the existence of a `ColorTheme` for that name.