Fix modifier content for NoticeModifier
We had some issues with the `.inlineNotice` modifier within the app due
to the `.clipped` modifier usage.

The main issue is that we were clipping the content itself in an attempt
to clip the overlay that animates in/out the notice view.

Then I reordered the modifiers to only apply `clipped` to the
`noticeOverlay` rather than the original content.

I tested this change in the iOS app and this fixes the issues we saw.