Add id to SATSButton spinner
I notice if we change a given button from a primary to a secondary style,
the loading indicator is not visible, this is because:

1. if the button was in `.secondary` style
2. we show the indicator (with a blue color on top)
3. we change the button to `.primary` style
4. we set the button to `loading`, but we don't see the indicator.

The indicator is still present, but not visible as the indicator style doesn't get
refreshed when changing the button's style.

Then the solution is to use the `style.name` as `id` to make the spinner refresh
when the style is changed.