Fix `PullToRefreshContainer` reload behavior
with the work done in #85, I made the `PullToRefreshContainer` API
compatible with the `.refreshable` modifier.

To do that, i had to move the `isLoading` state, from the caller to the
internal state of the component.

The function that is passed as `onReload` to `PullToRefreshContainer` is
marked as async.

I was calling `await onReload()` to set the `isLoading = false`
**after the async onReload returned**, **BUT!!** since we had code that
was not using functions that use the new `async` annotation with the
`.refreshable` modifier, but completion hanlders in some cases. We
cannot rely on `await` then.

That caused multiple network calls, when performing the pull to refresh
gesture.

Now. Instead of relying on the `onReload` returning synchronously,
`isLoading` is taking in account the gesture of pull to refresh.