Keep the selected date in focus when scrolling results.
Given the structure we have in the app

1. date component
2. search results for the selected page

You can swipe the pages in (2), then when the next page is out of the
screen for the calendar component (1), we will scroll so that date
becomes visible.

To achieve this we use a `ScrollViewReader` and set the ids of each
view in the calendar.

The id is set from a date in the format "yyyy-MM-dd" as we don't want
variations in the hour and minute made 2 dates within the same day
diffrent.

I also adjusted carefully the way we render the date cells, so when
calling `proxy.scroll(to:)` works correctly. Previously, this didn't
work when the padding was surrounding the scroll view content for
example.