Add `SATSCell.BasicCell`
This is an attempt to replicate what UIKit does with `UITableViewCell`.
As we can have basic cells with configurable views, chevrons and so on,
but this time in SwiftUI.

Even though writing UI for cells is quite easy, I made this to have a
common way to define cell-like buttons.

When using SwiftUI we tend not to use the `List` element, as they are
quite restrictive and we don't have concrete benefits most of the time
with it anyway.

Then creating a standard SwiftUI implementation will help us remove
extra code.

Code structure:
--------------

I was iterating a lot when it comes to the API we expose in this
component.

I didn't use a `CellViewData` as that structure would have to define
actual swiftUI views in its properties in a generic way, which didn't
feel correct.

The purpose of the namespace `SATSCell` is to be able to have a common
type to expand with default views, like `SATSCell.SystemChevron` etc.
We coulnd't simply extend `SATSCell.BasicCell` as it's a generic type.

The idea is that in `SATSCell` we can create commonly used views for the
cells, like the `SATSCell.ExternalIcon`.

Check the demo view for samples on how to use this code.

Divider:
-------

One of the aspects I was a bit on the fence about is that with this
implementation of the custom cell, we delegate the responsibility to the
caller to not include a separator in the last cell of a section, with
the `includeDivider` paramter. It feels a bit weird overall.

As always I had a bunch of experimentation that led to this design in a
way that it wouldn't go "too far" when it comes to try to generalize
everything for every use case.