Use `Font.custom` to implement the usage of custom fonts for SwiftUI
SATSFont.font internally uses `UIFontMetrics` to determine the sizes of
the font adjusted to font families for dynamic type.

**But** since this lives in UIKit-land, this value doens't get
recomputed when the environment changes (in a SwiftUI context), but only
when the UITraitCollection of a view controller changes.

In the case of SwiftUI previews, only the environment changes for
previews, that's the reason we were not seeing the changes in font size
when we did the "show accessibility variants" in previews.

Now, by using `Font.custom` the values do get recomputed when the
environment's `dynamicTypeSize` changes.