Re-define `Font.satsFont`
to accept `Font.Design` as a parameter, but also
to use the system font when the "weight" is not "SATS Feeling"

this is to instead of using a custom font, to use the system one.

the system one allows us to use the `design` parameter.

Since the `.system` font doesn't have a parameter like `Font.custom(**relativeTo**)`, we need to
manually scale the font size when dynamic type changes.

This is handled with the parameters.

Now, `Font.satsFont(_, weight:design:)` is still available, for compatibility purposes only, but the variant
with the `scaledSize` should be preferred instead.