`View.satsFont` uses the system font
Via an internal view modifier.

`Font.satsFont` should receive the parameter `scaledSize` to deal with dynamic type changes for the system font.

the current `dynamicTypeSize` is held in the environment for SwiftUI. That's the reason I needed a view modifier, so
i could use a `@ScaledMetric` to calculate the scaled font size we needed. This helps us to keep previews reacting
to dynamic size changes.