Remove `Text.satsFont() -> Text`
Since we want to have previews that react to dynamic type changes, i need to use a `ViewModifer` to set the font.

but `Text.modifier(some Modifier)` returns `some View`, not `Text` as the signature of this method.

Since we are mostly dealing with `Text.satsFont` in the context of a view, this is not a problem, although technically
we are changing the public API of SATSCore, making this a breaking change