Implement new color system
within the new color system there are some important definitions:

Color Primitive:
----------------

When defining a semantic color, previously we defined a color like
`action` by using the asset catalog to specify it's `light` and `dark`
values. One problem with that, is that we would replace the light value,
we most certaintly want to replace the value of all the asset catalog
colors that use the same HEX value.

Now, instead of using the Asset Catalog, we define **Color Primitives**
that are the actual color HEX values used in the app. A semantic color
then is defined based on `ColorPrimitive` for its light and dark values.

Since within the Figma and app communication, HEX colors are the most
useful, we define primitives with HEX colors and we also ensure we can
compute an actual `UIColor` from that hex value.

SATSColorName:
--------------

Used to define the semantic colors for the app.

SemanticColorDefinition:
-----------------------

Now, given a `SATSColorName`, like `.action`, the actual semantic color
definition in based of `ColorPrimitive` is captured by this value.

```
action: SemanticColorDefinition(light: .satsCoral100, dark: .satsCoral80)
```

within that `SemanticColorDefinition` we can extract the `uiColor` or
the `color`.

Notice that:

- for watchOS we only use the darkMode color values, there is no
  light/dark mode in this platform
- we use `uiColor` to define a `Color` as `UIColor` has the
  `dynamicProvider` initializer that allows to specify specific colors
  for light and dark mode.

Inspired by: https://www.jessesquires.com/blog/2023/07/11/creating-dynamic-colors-in-swiftui/

SATSColor Definitions:
----------------------

Instead of making the definitions like

```
action: SemanticColorDefinition(light: .satsCoral100, dark: .satsCoral80)
```

In both SwiftUI and UIKit constants, we just have a single source of
truth for those definitions, living in a dictionary under `SATSColor`.

UIKit and SwiftUI constants:
----------------------------

Each `SATSColorName` needs to be exposed for both `UIKit` and `SwiftUI`.

I made experiments were trying to use `@dynamicMemberLookup` to use the
`SATSColor.definitions` to dynamically create the constants for
`UIColor` and `Color`. The main drawback of this approach was that we
lose autocompletion, which is quite useful when using these constants.

I was also exploring posibilities to nest the SATSColor constants under
something like

```swift
Color.sats.action

// instead of

Color.action
```

So we could not mixed the system constants and our constants. This was
not something easy to do, so I opted to manually define the constants as
before.